feat: add search page
- refactor asset
- add algoliasearch
