+++ assets/js/iconfont.js
@@ -1,1 +1,1 @@
-!function(l){var e,n='<svg><symbol id="icon-close" viewBox="0 0 1024 1024"><path d="M556.8 512l265.6-265.6c12.8-12.8 12.8-32 0-44.8s-32-12.8-44.8 0L512 467.2 246.4 201.6c-12.8-12.8-32-12.8-44.8 0s-12.8 32 0 44.8l265.6 265.6-265.6 265.6c-12.8 12.8-12.8 32 0 44.8 6.4 6.4 12.8 9.6 22.4 9.6s16-3.2 22.4-9.6l265.6-265.6 265.6 265.6c6.4 6.4 16 9.6 22.4 9.6s16-3.2 22.4-9.6c12.8-12.8 12.8-32 0-44.8L556.8 512z"  ></path></symbol><symbol id="icon-menu" viewBox="0 0 1024 1024"><path d="M128 298.666667h768a42.666667 42.666667 0 0 0 0-85.333334H128a42.666667 42.666667 0 0 0 0 85.333334z m768 170.666666H128a42.666667 42.666667 0 0 0 0 85.333334h768a42.666667 42.666667 0 0 0 0-85.333334z m0 256H128a42.666667 42.666667 0 0 0 0 85.333334h768a42.666667 42.666667 0 0 0 0-85.333334z"  ></path></symbol><symbol id="icon-arrow" viewBox="0 0 1024 1024"><path d="M318.73024 836.32128a20.41856 20.41856 0 0 0 28.95872 0l307.2-307.2a20.45952 20.45952 0 0 0 0-28.95872l-307.2-307.2a20.45952 20.45952 0 1 0-28.95872 28.95872l292.72064 292.72064-292.72064 292.72064a20.45952 20.45952 0 0 0 0 28.95872z" fill="#231F20" ></path></symbol><symbol id="icon-copy" viewBox="0 0 1024 1024"><path d="M877.714286 0H265.142857c-5.028571 0-9.142857 4.114286-9.142857 9.142857v64c0 5.028571 4.114286 9.142857 9.142857 9.142857h566.857143v786.285715c0 5.028571 4.114286 9.142857 9.142857 9.142857h64c5.028571 0 9.142857-4.114286 9.142857-9.142857V36.571429c0-20.228571-16.342857-36.571429-36.571428-36.571429zM731.428571 146.285714H146.285714c-20.228571 0-36.571429 16.342857-36.571428 36.571429v606.514286c0 9.714286 3.885714 18.971429 10.742857 25.828571l198.057143 198.057143c2.514286 2.514286 5.371429 4.571429 8.457143 6.285714v2.171429h4.8c4 1.485714 8.228571 2.285714 12.571428 2.285714H731.428571c20.228571 0 36.571429-16.342857 36.571429-36.571429V182.857143c0-20.228571-16.342857-36.571429-36.571429-36.571429zM363.428571 950.857143h-0.228571L192 779.657143v-0.228572h171.428571v171.428572z"  ></path></symbol><symbol id="icon-transfer" viewBox="0 0 1024 1024"><path d="M769.841131 507.345455a31.340606 31.340606 0 0 0-44.187151 0 30.947556 30.947556 0 0 0 0 43.959596l179.613737 178.424242H43.618263a31.030303 31.030303 0 1 0 0 62.060606h861.390868L725.385051 970.472727a30.947556 30.947556 0 0 0 0 43.959596c6.237091 6.206061 14.03604 9.050505 22.093575 9.050505 8.067879 0 16.115071-3.10303 22.093576-9.050505l232.892768-231.692929a30.616566 30.616566 0 0 0 9.102222-21.979798c0-8.274747-3.382303-16.290909-9.102222-21.979798L769.861818 507.345455zM254.158869 516.654545c6.237091 6.206061 14.025697 9.050505 22.093575 9.050506 8.057535 0 16.115071-3.10303 22.093576-9.050506a30.947556 30.947556 0 0 0 0-43.959596L118.732283 294.012121h861.401212a31.030303 31.030303 0 0 0 31.185454-31.030303c0-17.066667-14.03604-31.030303-31.195797-31.030303H118.990869L298.614949 53.268687a30.947556 30.947556 0 0 0 0-43.959596 31.340606 31.340606 0 0 0-44.187151 0L21.514343 241.00202a30.947556 30.947556 0 0 0 0 43.959596l232.623839 231.692929z" fill="#2E323F" ></path></symbol><symbol id="icon-toc-arrow" viewBox="0 0 1024 1024"><path d="M20.48 245.76h983.04L512 778.24z"  ></path></symbol></svg>',t=(e=document.getElementsByTagName("script"))[e.length-1].getAttribute("data-injectcss");if(t&&!l.__iconfont__svg__cssinject__){l.__iconfont__svg__cssinject__=!0;try{document.write("<style>.svgfont {display: inline-block;width: 1em;height: 1em;fill: currentColor;vertical-align: -0.1em;font-size:16px;}</style>")}catch(e){console&&console.log(e)}}!function(e){if(document.addEventListener)if(~["complete","loaded","interactive"].indexOf(document.readyState))setTimeout(e,0);else{var t=function(){document.removeEventListener("DOMContentLoaded",t,!1),e()};document.addEventListener("DOMContentLoaded",t,!1)}else document.attachEvent&&(o=e,i=l.document,c=!1,(a=function(){try{i.documentElement.doScroll("left")}catch(e){return void setTimeout(a,50)}n()})(),i.onreadystatechange=function(){"complete"==i.readyState&&(i.onreadystatechange=null,n())});function n(){c||(c=!0,o())}var o,i,c,a}(function(){var e,t;(e=document.createElement("div")).innerHTML=n,n=null,(t=e.getElementsByTagName("svg")[0])&&(t.setAttribute("aria-hidden","true"),t.style.position="absolute",t.style.width=0,t.style.height=0,t.style.overflow="hidden",function(e,t){t.firstChild?function(e,t){t.parentNode.insertBefore(e,t)}(e,t.firstChild):t.appendChild(e)}(t,document.body))})}(window);
+!function(l){var e,o='<svg><symbol id="icon-search" viewBox="0 0 1024 1024"><path d="M927.104 866.816l-195.626667-208.768C780.629333 601.770667 810.666667 528.426667 810.666667 448 810.666667 271.530667 667.136 128 490.666667 128S170.666667 271.530667 170.666667 448 314.197333 768 490.666667 768c65.322667 0 126.08-19.754667 176.768-53.461333l197.461333 210.688L927.104 866.816zM256 448C256 318.592 361.258667 213.333333 490.666667 213.333333S725.333333 318.592 725.333333 448 620.074667 682.666667 490.666667 682.666667 256 577.408 256 448z"  ></path></symbol><symbol id="icon-close" viewBox="0 0 1024 1024"><path d="M556.8 512l265.6-265.6c12.8-12.8 12.8-32 0-44.8s-32-12.8-44.8 0L512 467.2 246.4 201.6c-12.8-12.8-32-12.8-44.8 0s-12.8 32 0 44.8l265.6 265.6-265.6 265.6c-12.8 12.8-12.8 32 0 44.8 6.4 6.4 12.8 9.6 22.4 9.6s16-3.2 22.4-9.6l265.6-265.6 265.6 265.6c6.4 6.4 16 9.6 22.4 9.6s16-3.2 22.4-9.6c12.8-12.8 12.8-32 0-44.8L556.8 512z"  ></path></symbol><symbol id="icon-menu" viewBox="0 0 1024 1024"><path d="M128 298.666667h768a42.666667 42.666667 0 0 0 0-85.333334H128a42.666667 42.666667 0 0 0 0 85.333334z m768 170.666666H128a42.666667 42.666667 0 0 0 0 85.333334h768a42.666667 42.666667 0 0 0 0-85.333334z m0 256H128a42.666667 42.666667 0 0 0 0 85.333334h768a42.666667 42.666667 0 0 0 0-85.333334z"  ></path></symbol><symbol id="icon-ARROW" viewBox="0 0 1024 1024"><path d="M318.73024 836.32128a20.41856 20.41856 0 0 0 28.95872 0l307.2-307.2a20.45952 20.45952 0 0 0 0-28.95872l-307.2-307.2a20.45952 20.45952 0 1 0-28.95872 28.95872l292.72064 292.72064-292.72064 292.72064a20.45952 20.45952 0 0 0 0 28.95872z" fill="#231F20" ></path></symbol><symbol id="icon-copy" viewBox="0 0 1024 1024"><path d="M877.714286 0H265.142857c-5.028571 0-9.142857 4.114286-9.142857 9.142857v64c0 5.028571 4.114286 9.142857 9.142857 9.142857h566.857143v786.285715c0 5.028571 4.114286 9.142857 9.142857 9.142857h64c5.028571 0 9.142857-4.114286 9.142857-9.142857V36.571429c0-20.228571-16.342857-36.571429-36.571428-36.571429zM731.428571 146.285714H146.285714c-20.228571 0-36.571429 16.342857-36.571428 36.571429v606.514286c0 9.714286 3.885714 18.971429 10.742857 25.828571l198.057143 198.057143c2.514286 2.514286 5.371429 4.571429 8.457143 6.285714v2.171429h4.8c4 1.485714 8.228571 2.285714 12.571428 2.285714H731.428571c20.228571 0 36.571429-16.342857 36.571429-36.571429V182.857143c0-20.228571-16.342857-36.571429-36.571429-36.571429zM363.428571 950.857143h-0.228571L192 779.657143v-0.228572h171.428571v171.428572z"  ></path></symbol><symbol id="icon-transfer" viewBox="0 0 1024 1024"><path d="M769.841131 507.345455a31.340606 31.340606 0 0 0-44.187151 0 30.947556 30.947556 0 0 0 0 43.959596l179.613737 178.424242H43.618263a31.030303 31.030303 0 1 0 0 62.060606h861.390868L725.385051 970.472727a30.947556 30.947556 0 0 0 0 43.959596c6.237091 6.206061 14.03604 9.050505 22.093575 9.050505 8.067879 0 16.115071-3.10303 22.093576-9.050505l232.892768-231.692929a30.616566 30.616566 0 0 0 9.102222-21.979798c0-8.274747-3.382303-16.290909-9.102222-21.979798L769.861818 507.345455zM254.158869 516.654545c6.237091 6.206061 14.025697 9.050505 22.093575 9.050506 8.057535 0 16.115071-3.10303 22.093576-9.050506a30.947556 30.947556 0 0 0 0-43.959596L118.732283 294.012121h861.401212a31.030303 31.030303 0 0 0 31.185454-31.030303c0-17.066667-14.03604-31.030303-31.195797-31.030303H118.990869L298.614949 53.268687a30.947556 30.947556 0 0 0 0-43.959596 31.340606 31.340606 0 0 0-44.187151 0L21.514343 241.00202a30.947556 30.947556 0 0 0 0 43.959596l232.623839 231.692929z" fill="#2E323F" ></path></symbol><symbol id="icon-arrow" viewBox="0 0 1024 1024"><path d="M20.48 245.76h983.04L512 778.24z"  ></path></symbol></svg>',t=(e=document.getElementsByTagName("script"))[e.length-1].getAttribute("data-injectcss");if(t&&!l.__iconfont__svg__cssinject__){l.__iconfont__svg__cssinject__=!0;try{document.write("<style>.svgfont {display: inline-block;width: 1em;height: 1em;fill: currentColor;vertical-align: -0.1em;font-size:16px;}</style>")}catch(e){console&&console.log(e)}}!function(e){if(document.addEventListener)if(~["complete","loaded","interactive"].indexOf(document.readyState))setTimeout(e,0);else{var t=function(){document.removeEventListener("DOMContentLoaded",t,!1),e()};document.addEventListener("DOMContentLoaded",t,!1)}else document.attachEvent&&(n=e,i=l.document,c=!1,(a=function(){try{i.documentElement.doScroll("left")}catch(e){return void setTimeout(a,50)}o()})(),i.onreadystatechange=function(){"complete"==i.readyState&&(i.onreadystatechange=null,o())});function o(){c||(c=!0,n())}var n,i,c,a}(function(){var e,t;(e=document.createElement("div")).innerHTML=o,o=null,(t=e.getElementsByTagName("svg")[0])&&(t.setAttribute("aria-hidden","true"),t.style.position="absolute",t.style.width=0,t.style.height=0,t.style.overflow="hidden",function(e,t){t.firstChild?function(e,t){t.parentNode.insertBefore(e,t)}(e,t.firstChild):t.appendChild(e)}(t,document.body))})}(window);
