+++ content/en/search/_index.md
@@ -1,0 +1,3 @@
+---
+layout: "search"
+---
