+++ layouts/partials/header.html
@@ -19,7 +19,7 @@
               <span>
                 {{ .Name }}
               </span>
-              <svg class="icon" aria-hidden="true"><use xlink:href="#icon-arrow"></use></svg>
+              <svg class="icon" aria-hidden="true"><use xlink:href="#icon-ARROW"></use></svg>
             </a>
 					{{ end }}
 					{{ if .IsTranslated }}
