+++ layouts/partials/toc/toc_leaf.html
@@ -18,7 +18,7 @@
 			<li class="item">
 				<div class="link">
 					<svg class="icon arrow" aria-hidden="true">
-						<use xlink:href="#icon-toc-arrow"></use>
+						<use xlink:href="#icon-arrow"></use>
 					</svg>
 					<a
 						title="{{ .title }}"
